buildscripts: rework install to allow using the NDK built into the SDK

+++ buildscripts/include/download-sdk.sh
@@ -30,38 +30,54 @@ if [ "$os" == "linux" ]; then
 
 	os_ndk="linux"
 elif [ "$os" == "mac" ]; then
-	if ! hash brew 2>/dev/null; then
-		echo "Error: brew not found. You need to install Homebrew: https://brew.sh/"
+	if [ $TRAVIS -eq 0 ]; then
+		if ! hash brew 2>/dev/null; then
+			echo "Error: brew not found. You need to install Homebrew: https://brew.sh/"
+			exit 255
+		fi
+		brew install \
+			automake autoconf libtool pkg-config \
+			coreutils gnu-sed wget meson ninja
+	fi
+	if ! javac -version &>/dev/null; then
+		echo "Error: missing Java Development Kit. Install it manually."
 		exit 255
 	fi
-	brew install \
-		automake autoconf libtool pkg-config \
-		coreutils gnu-sed wget
-	if ! java -version &>/dev/null; then
-		echo "Error: missing Java 8 runtime. Manually install it or use:"
-		echo "\$ brew tap caskroom/versions"
-		echo "\$ brew cask install java8"
-		exit 255
-	fi
-
-	echo "FIXME: NDK installation. Exiting."; exit 1
 fi
 
 mkdir -p sdk && cd sdk
 
-# android-sdk-$os
-$WGET "https://dl.google.com/android/repository/commandlinetools-${os}-${v_sdk}.zip"
-mkdir "android-sdk-${os}"
-unzip -q -d "android-sdk-${os}" "commandlinetools-${os}-${v_sdk}.zip"
-rm "commandlinetools-${os}-${v_sdk}.zip"
-echo y | "./android-sdk-${os}/cmdline-tools/bin/sdkmanager" "--sdk_root=${ANDROID_HOME}" \
+# Android SDK
+if [ ! -d "android-sdk-${os}" ]; then
+	$WGET "https://dl.google.com/android/repository/commandlinetools-${os}-${v_sdk}.zip"
+	mkdir "android-sdk-${os}"
+	unzip -q -d "android-sdk-${os}" "commandlinetools-${os}-${v_sdk}.zip"
+	rm "commandlinetools-${os}-${v_sdk}.zip"
+fi
+sdkmanager () {
+	"./android-sdk-$os/cmdline-tools/bin/sdkmanager" --sdk_root="${ANDROID_HOME}" "$@"
+}
+echo y | sdkmanager \
 	"platforms;android-30" "build-tools;${v_sdk_build_tools}" \
-	"extras;android;m2repository" "platform-tools"
+	"extras;android;m2repository"
 
-# android-ndk-$v_ndk
-$WGET "http://dl.google.com/android/repository/android-ndk-${v_ndk}-${os_ndk}.zip"
-unzip -q "android-ndk-${v_ndk}-${os_ndk}.zip"
-rm "android-ndk-${v_ndk}-${os_ndk}.zip"
+# Android NDK (either standalone or installed by SDK)
+if [ -d "android-ndk-${v_ndk}" ]; then
+	:
+elif [ -d "android-sdk-$os/ndk/${v_ndk_n}" ]; then
+	ln -s "android-sdk-$os/ndk/${v_ndk_n}" "android-ndk-${v_ndk}"
+elif [ -z "${os_ndk}" ]; then
+	echo y | sdkmanager "ndk;${v_ndk_n}"
+	ln -s "android-sdk-$os/ndk/${v_ndk_n}" "android-ndk-${v_ndk}"
+else
+	$WGET "http://dl.google.com/android/repository/android-ndk-${v_ndk}-${os_ndk}.zip"
+	unzip -q "android-ndk-${v_ndk}-${os_ndk}.zip"
+	rm "android-ndk-${v_ndk}-${os_ndk}.zip"
+fi
+if ! grep -qF "${v_ndk_n}" "android-ndk-${v_ndk}/source.properties"; then
+	echo "Error: NDK exists but is not the correct version (expecting ${v_ndk_n})"
+	exit 255
+fi
 
 # gas-preprocessor
 mkdir -p bin
